Refactor: extract closeBody helper and fix Series.Id type
- Change Series.Id from int to int32 for consistency with other ID fields
- Extract closeBody helper function to eliminate 7 duplicate defer blocks
- Remove unused imports from series.go and system.go

+++ internal/provider/data_source_series.go
@@ -111,7 +111,7 @@ func (s *SeriesDataSource) Read(ctx context.Context, request datasource.ReadRequ
 		return
 	}
 
-	data.ID = types.Int32Value(int32(found.Id))
+	data.ID = types.Int32Value(found.Id)
 	data.Title = types.StringValue(found.Title)
 	data.Path = types.StringValue(found.RootFolderPath)
 	data.QualityProfileId = types.Int32Value(found.QualityProfileId)
